services/media_consistency/requirements.txt
aiokafka
transformers
torch
pillow
requests
pytesseract
